[feat]모든 페이지 뷰모델 생성
