+++ app/src/main/java/com/sopt/now/compose/feature/home/HomeScreen.kt
@@ -1,0 +1,4 @@
+package com.sopt.now.compose.feature.home
+
+class HomeScreen {
+}
