+++ app/src/main/java/com/sopt/now/compose/feature/home/HomeViewModel.kt
@@ -1,0 +1,7 @@
+package com.sopt.now.compose.feature.home
+
+import androidx.lifecycle.ViewModel
+
+class HomeViewModel : ViewModel() {
+
+}
