+++ app/src/main/java/com/sopt/now/compose/model/BottomNavigationItem.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose
+package com.sopt.now.compose.model
 
 import androidx.compose.ui.graphics.vector.ImageVector
