+++ app/src/main/java/com/sopt/now/compose/remote/response/ResponseSignUpDto.kt
@@ -1,10 +1,10 @@
-package com.sopt.now.compose.response
+package com.sopt.now.compose.remote.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ResponseChangePwdDto (
+data class ResponseSignUpDto (
     @SerialName("code")
     val code: Int,
     @SerialName("message")
