+++ app/src/main/java/com/sopt/now/compose/remote/service/AuthService.kt
@@ -1,9 +1,9 @@
-package com.sopt.now.compose.service
+package com.sopt.now.compose.remote.service
 
-import com.sopt.now.compose.request.RequestLoginDto
-import com.sopt.now.compose.request.RequestSignUpDto
-import com.sopt.now.compose.response.ResponseLoginDto
-import com.sopt.now.compose.response.ResponseSignUpDto
+import com.sopt.now.compose.remote.request.RequestLoginDto
+import com.sopt.now.compose.remote.request.RequestSignUpDto
+import com.sopt.now.compose.remote.response.ResponseLoginDto
+import com.sopt.now.compose.remote.response.ResponseSignUpDto
 import retrofit2.Call
 import retrofit2.http.Body
 import retrofit2.http.POST
